+++ internal/service/golden_gate/register_datasource.go
@@ -14,4 +14,10 @@ func RegisterDatasource() {
 	tfresource.RegisterDatasource("oci_golden_gate_deployment_upgrade", GoldenGateDeploymentUpgradeDataSource())
 	tfresource.RegisterDatasource("oci_golden_gate_deployment_upgrades", GoldenGateDeploymentUpgradesDataSource())
 	tfresource.RegisterDatasource("oci_golden_gate_deployments", GoldenGateDeploymentsDataSource())
+	tfresource.RegisterDatasource("oci_golden_gate_message", GoldenGateMessageDataSource())
+	tfresource.RegisterDatasource("oci_golden_gate_messages", GoldenGateMessagesDataSource())
+	tfresource.RegisterDatasource("oci_golden_gate_trail_file", GoldenGateTrailFileDataSource())
+	tfresource.RegisterDatasource("oci_golden_gate_trail_files", GoldenGateTrailFilesDataSource())
+	tfresource.RegisterDatasource("oci_golden_gate_trail_sequence", GoldenGateTrailSequenceDataSource())
+	tfresource.RegisterDatasource("oci_golden_gate_trail_sequences", GoldenGateTrailSequencesDataSource())
 }