+++ internal/service/log_analytics/register_resource.go
@@ -13,6 +13,8 @@ func RegisterResource() {
 	tfresource.RegisterResource("oci_log_analytics_log_analytics_preferences_management", LogAnalyticsLogAnalyticsPreferencesManagementResource())
 	tfresource.RegisterResource("oci_log_analytics_log_analytics_resource_categories_management", LogAnalyticsLogAnalyticsResourceCategoriesManagementResource())
 	tfresource.RegisterResource("oci_log_analytics_log_analytics_unprocessed_data_bucket_management", LogAnalyticsLogAnalyticsUnprocessedDataBucketManagementResource())
+	tfresource.RegisterResource("oci_log_analytics_namespace_ingest_time_rule", LogAnalyticsNamespaceIngestTimeRuleResource())
+	tfresource.RegisterResource("oci_log_analytics_namespace_ingest_time_rules_management", LogAnalyticsNamespaceIngestTimeRulesManagementResource())
 	tfresource.RegisterResource("oci_log_analytics_namespace_scheduled_task", LogAnalyticsNamespaceScheduledTaskResource())
 	tfresource.RegisterResource("oci_log_analytics_namespace", LogAnalyticsNamespaceResource())
 }